rolevate-agent/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# AI & LLM
langchain==0.1.4
langchain-openai==0.0.5
openai==1.10.0
tiktoken==0.5.2

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-docx-template==0.16.7
weasyprint==60.2
Jinja2==3.1.3

# File Handling
python-magic==0.4.27
aiofiles==23.2.1

# Utils
python-dotenv==1.0.0
loguru==0.7.2
click==8.1.7

# Optional: Database
sqlalchemy==2.0.25
alembic==1.13.1

# Optional: Storage
boto3==1.34.34

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14
